User Roles and Permissions Protection

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Models\Post;
 use Inertia\Inertia;
 use Inertia\Response;
 use App\Http\Resources\PostResource;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\RedirectResponse;
 use App\Http\Requests\StorePostRequest;
 
@@ -13,7 +14,7 @@ class PostController extends Controller
 {
     public function index(): Response
     {
-//        sleep(3);
+        Gate::authorize('viewAny', Post::class);
 
         $posts = PostResource::collection(Post::all());
 
@@ -22,12 +23,15 @@ class PostController extends Controller
 
     public function create(): Response
     {
+        Gate::authorize('create', Post::class);
+
         return Inertia::render('Posts/Create');
     }
 
     public function store(StorePostRequest $request): RedirectResponse
     {
-//        sleep(2);
+        Gate::authorize('create', Post::class);
+
         Post::create($request->validated());
 
         return redirect()->route('posts.index')
@@ -36,11 +40,15 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
+        Gate::authorize('create', Post::class);
+
         return Inertia::render('Posts/Edit', compact('post'));
     }
 
     public function update(Post $post, StorePostRequest $request)
     {
+        Gate::authorize('create', Post::class);
+
         $post->update($request->validated());
 
         return redirect()->route('posts.index')
@@ -49,6 +57,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        Gate::authorize('create', Post::class);
+
         $post->delete();
 
         return redirect()->route('posts.index')
